# src/fk-bff-gerente/openapi.yaml
openapi: 3.1.0
info:
  title: fk-bff-gerente
  description: "Backend-for-frontend do GerenteRelacionamento. Fronteira web -- nao\
    \ e bounded context (CONTEXT-MAP.md). Conduz Authorization Code + PKCE + OIDC\
    \ contra fk-servidor-autorizacao, mantem a sessao em Redis via Spring Session\
    \ e expoe ao app-gerente somente o que o Angular precisa: identidade da sessao\
    \ corrente, a carteira, e a composicao da tela de atendimento a partir de fk-servico-carteira-clientes\
    \ e fk-servico-credito (AC30). Nenhum access token ou refresh token atravessa\
    \ esta fronteira para o browser (AC19) -- o BFF obtem por Token Exchange um token\
    \ para cada destino e nunca reencaminha o token de login (ADR-0015). Nao implementa\
    \ regra de credito, nao fala com fk-simulador-core-legado e nao substitui a autorizacao\
    \ de recurso feita pelos servicos. Produz o seu PROPRIO envelope publico de erro\
    \ -- nunca repassa o ProblemDetail upstream verbatim."
  version: "0003"
servers:
- url: /bff
  description: "Origem publica atras do nginx (docker compose), TLS local."
...
paths:
  /api/clientes/{clienteId}/contas/{contaId}/solicitacoes-aumento-limite:
    post:
      tags:
      - solicitacoes
      summary: Proxy autenticado da submissao da SolicitacaoAumentoLimite
      description: "Encaminhamento puro para fk-servico-credito, com token trocado\
        \ por Token Exchange (credito.escrita -- registration distinta da usada no\
        \ GET do limite vigente, least privilege por operacao). Corpo e Idempotency-Key\
        \ atravessam INTACTOS: o BFF nao gera, nao regenera e nao reinterpreta nenhum\
        \ dos dois. Status HTTP de sucesso (201/200) e o corpo sao propagados tal\
        \ como fk-servico-credito devolveu. Escrita real, sujeita a CSRF (AC20)."
      operationId: submeterSolicitacaoAumentoLimite
      parameters:
      - name: clienteId
        in: path
        description: Identificador do Cliente selecionado.
        required: true
        schema:
          type: string
          pattern: "^[0-9]{1,10}$"
        example: 1
      - name: contaId
        in: path
        description: Identificador da ContaCorrente.
        required: true
        schema:
          type: string
          pattern: "^[0-9]{1,10}$"
        example: 10001
      - name: Idempotency-Key
        in: header
        description: Repassado intacto para fk-servico-credito -- nunca gerado ou
          reinterpretado aqui.
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        description: Repassado intacto para fk-servico-credito -- o BFF nao desserializa
          este corpo.
        content:
          application/json:
            schema:
              type: object
              description: "Corpo cru, nao desserializado no BFF."
        required: true
      responses:
        "201":
          description: SolicitacaoAumentoLimite criada agora (independentemente do
            resultado da decisao).
          content:
            application/json:
              schema:
                type: object
                description: "Corpo cru propagado de fk-servico-credito, sem tipagem\
                  \ no BFF."
        "200":
          description: "Replay idempotente -- mesma chave, mesmo fingerprint, operacao\
            \ ja concluida."
          content:
            application/json:
              schema:
                type: object
                description: "Corpo cru propagado de fk-servico-credito, sem tipagem\
                  \ no BFF."
        "400":
          description: "IDEMPOTENCY_KEY_AUSENTE (header ausente, recusado pelo proprio\
            \ BFF), IDEMPOTENCY_KEY_INVALIDA ou COMANDO_ILEGIVEL (fk-servico-credito),\
            \ ou IDENTIFICADOR_INVALIDO (clienteId/contaId fora do formato, recusado\
            \ pelo proprio BFF)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EnvelopeErroPublico"
        "401":
          description: Sem sessao autenticada.
        "403":
          description: Sem direito de atendimento atual. codigo = SEM_DIREITO_DE_ATENDIMENTO.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EnvelopeErroPublico"
        "404":
          description: A conta nao e reconhecida pelo CoreLegado. codigo = CONTA_NAO_ENCONTRADA.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EnvelopeErroPublico"
        "409":
          description: "LIMITE_VIGENTE_DESATUALIZADO, SOLICITACAO_NAO_TERMINAL_EXISTENTE\
            \ ou IDEMPOTENCIA_EM_PROCESSAMENTO."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EnvelopeErroPublico"
        "422":
          description: "COMANDO_INVALIDO, LIMITE_SOLICITADO_NAO_AUMENTA ou IDEMPOTENCIA_FINGERPRINT_DIVERGENTE."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EnvelopeErroPublico"
        "502":
          description: "codigo desconhecido, corpo upstream ilegivel, ou token delegado\
            \ recusado (401 da cadeia de Token Exchange -- nunca vira 401 para o browser).\
            \ codigo = DEPENDENCIA_INDISPONIVEL."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EnvelopeErroPublico"
        "503":
          description: fk-servico-credito indisponivel. codigo = DEPENDENCIA_INDISPONIVEL.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EnvelopeErroPublico"
      security:
      - cookieSessao: []
...
components:
  schemas:
    EnvelopeErroPublico:
      type: object
      description: "Envelope publico de erro do BFF -- so status+codigo, nunca o ProblemDetail\
        \ (RFC 7807) de um Resource Server verbatim."
      properties:
        status:
          type: integer
          format: int32
          description: "Status HTTP, espelhado do valor devolvido na resposta."
          example: 403
        codigo:
          type: string
          description: Codigo estavel do dominio. Omitido (nao null explicito) quando
            o upstream nao publica um -- caso de fk-servico-carteira-clientes em 403/404.
          enum:
          - IDEMPOTENCY_KEY_AUSENTE
          - IDEMPOTENCY_KEY_INVALIDA
          - COMANDO_ILEGIVEL
          - IDENTIFICADOR_INVALIDO
          - SEM_DIREITO_DE_ATENDIMENTO
          - CONTA_NAO_ENCONTRADA
          - LIMITE_VIGENTE_DESATUALIZADO
          - SOLICITACAO_NAO_TERMINAL_EXISTENTE
          - IDEMPOTENCIA_EM_PROCESSAMENTO
          - COMANDO_INVALIDO
          - LIMITE_SOLICITADO_NAO_AUMENTA
          - IDEMPOTENCIA_FINGERPRINT_DIVERGENTE
          - DEPENDENCIA_INDISPONIVEL
...
  securitySchemes:
    cookieSessao:
      type: apiKey
      description: "Cookie de sessao opaco (Spring Session/Redis), HttpOnly, Secure,\
        \ SameSite=Lax. Nunca carrega access token ou refresh token -- eles ficam\
        \ no OAuth2AuthorizedClientRepository, server-side."
      name: SESSION
      in: cookie